blindspot-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
bcrypt==4.1.3
PyJWT==2.9.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.15.0
gTTS==2.5.1
faster-whisper==1.0.3